init: init auth modal

+++ src/components/Header/Header.jsx
@@ -2,15 +2,21 @@ import { NavLink } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 
 import css from './Header.module.css';
-import { removeUser } from 'redux/features/authSlice';
+import { removeUser, toggleModal } from 'redux/features/authSlice';
+import { clearLibraryMovies } from 'redux/features/librarySlice';
+
 import { SearchField } from 'components';
 
 export const Header = () => {
   const dispatch = useDispatch();
   const isAuth = useSelector(state => state.auth.isAuth);
 
-  const handleLogOut = () => {
-    dispatch(removeUser());
+  const handleClickLogBtn = () => {
+    if (isAuth) {
+      dispatch(removeUser());
+      dispatch(clearLibraryMovies());
+    }
+    dispatch(toggleModal(true));
   };
 
   return (
@@ -25,14 +31,6 @@ export const Header = () => {
               Home
             </NavLink>
           </li>
-          {/* <li className={css.navigationItem}>
-            <NavLink
-              className={({ isActive }) => (isActive ? css.active : undefined)}
-              to="/movies"
-            >
-              Movies
-            </NavLink>
-          </li> */}
           <li className={css.navigationItem}>
             <NavLink
               className={({ isActive }) => (isActive ? css.active : undefined)}
@@ -44,11 +42,9 @@ export const Header = () => {
         </ul>
       </nav>
       <SearchField />
-      {isAuth && (
-        <button className={css.logBtn} type="button" onClick={handleLogOut}>
-          Log out
-        </button>
-      )}
+      <button className={css.logBtn} type="button" onClick={handleClickLogBtn}>
+        {isAuth ? 'Logout' : 'Login'}
+      </button>
     </header>
   );
 };
